refactor(archive): move stats and coverage out of lib
