+++ docs/topology-status.md
@@ -29,9 +29,9 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 579
+- **Stable** (no move scoped): 581
 - **Kernel** (polylogue/ root): 10
-- **Primitives** (lib-root + storage-root): 21
+- **Primitives** (lib-root + storage-root): 19
 - **TBD** (cell needs explicit assignment): 0
 - **Total declared**: 697
 - **Realized polylogue/**/*.py**: 697 files declared
